create a new folder for integration tests
