+++ Makefile
@@ -1,11 +1,11 @@
 export PYTHONPATH=$(shell pwd)/hyperon_das_atomdb/
 export PYTHONDONTWRITEBYTECODE=1
 
-test:
-	@pytest -v
+test-integration:
+	@pytest -sx ./tests/integration
 
 test-coverage:
-	@py.test -sx ./tests --cov=./hyperon_das_atomdb/ --cov-report=term-missing --cov-fail-under=70
+	@py.test -sx ./tests/unit --cov=./hyperon_das_atomdb/ --cov-report=term-missing --cov-fail-under=70
 
 isort:
 	@isort ./hyperon_das_atomdb ./tests --multi-line=3 --trailing-comma --force-grid-wrap=0 --use-parentheses --line-width=79
