+++ tests/unit/adapters/test_ram_only.py
