+++ tests/unit/adapters/test_redis_mongo_db.py
